Add about sections, rename delegate to portfolio delegate and add blogs to it!
It'll be nice to have one central location for all of the data used. Changed up the header to look a bit more professional and to make it more obvious people have to scroll.

Created the AboutFooter which will now link to the latest blog post. Will get one of the 5 latest later.

go to sleep.